+++ Editor/FolderTabsSettings.cs
@@ -1,0 +1,15 @@
+using UnityEditor;
+
+namespace FolderTabs
+{
+    static class FolderTabsSettings
+    {
+        const string KeyInvertScroll = "FolderTabs.InvertScroll";
+
+        public static bool InvertScroll
+        {
+            get => EditorPrefs.GetBool(KeyInvertScroll, false);
+            set => EditorPrefs.SetBool(KeyInvertScroll, value);
+        }
+    }
+}